+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.26.1",
     "bootstrap": "^5.2.0-beta1",
+    "isexe": "^2.0.0",
     "moment": "^2.29.3",
     "react": "^17.0.2",
     "react-chartjs-2": "^4.1.0",
