+++ src/components/History/Tipos/Chart.jsx
